change button view - lile in design
