+++ kpoint-react/src/components/project-page/SubscribeButton.tsx
@@ -40,9 +40,20 @@ const SubscribeButton: FC<SubscribeButtonProps> = ({
 
   return (
     <Button
+      sx={{
+        border: '2px solid rgb(130, 130, 130)',
+        borderRadius: '5px',
+        background: 'rgb(255, 255, 255, 0)',
+        width: '260px',
+        height: '46px',
+        color: 'rgb(130, 130, 130)',
+        fontSize: '14px',
+        fontWeight: 500,
+        lineHeight: '100%',
+        letterSpacing: '0.5px',
+      }}
       size="small"
       startIcon={<ControlPointTwoToneIcon />}
-      sx={{ justifyContent: 'right' }}
       onClick={handleButtonSubClick}
     >
       {isFollowed ? t('buttons.unfollow') : t('buttons.follow')}
